Add "copy to clipboard" button for exports
I've written a ClipboardUtils.gd class that uses OS.execute to call a bunch of clipboard utils.
Web implementation was easiest(-ish)
Still has to be worked on a bit, and I haven't tested if the MacOS approach works as I do not have a Mac and I've failed trying to install Hackintosh.