Add better exporting/saving for Chromium-based browsers
This uses the new web filesystem access APis, making GodSVG no longer save a new file to the downloads folder.
This instead prompts the user once to save a file and stores a handle to that file that we can later write to all we want.
There are some limitations; Mainly, the handle for each file has to be requested again any time the user loads up GodSVG again, but this isn't that problematic and right now I'm doing it any time the user saves the file again.

See: https://developer.chrome.com/docs/capabilities/web-apis/file-system-access